[SNAKE-35] Added Level3 - Added matrix of bool for Level 3 Data - Implemented logic for Initialization for Level3 - Fixed type conversion

+++ source/app/Game/Game.h
@@ -23,7 +23,7 @@ public:
 	void Render();
 	bool IsGameOver();
 	bool IsLevelCompleted();
-	int GetAmountFoodLeft();
+	size_t GetAmountFoodLeft();
 	virtual ~Game();
 	size_t FoodLeft;
 protected:
